# RavenDB-24942 use Avx512F.VL if possible for the inner loop of XXhash