### Issue link

https://issues.hibernatingrhinos.com/issue/RavenDB-24942/Vectorize-XXhash64

### Additional description

This PR vectorizes the inner loop of `XXHash64` if `Avx512F.VL` is supported (meaning .NET 8 and above and a proper CPU). `VL` allows for various length rotations that makes the inner part of the loop a fold of a few vector operations. Looks like a potential 25% gain.

#### Before

| Method                    | Mean     | Error   | StdDev  | Ratio | Code Size |
|-------------------------- |---------:|--------:|--------:|------:|----------:|
| XXHash64_RawPointer       | 426.9 ns | 4.99 ns | 4.67 ns |  0.70 |     827 B |
| XXHash64_FixedPointer     | 424.9 ns | 5.57 ns | 5.21 ns |  0.70 |     880 B |
| XXHash64_FixedSpan        | 422.7 ns | 2.10 ns | 1.86 ns |  0.69 |     917 B |
| XXHash64_Span             | 419.2 ns | 0.74 ns | 0.57 ns |  0.69 |     840 B |
| XXHash64_SpanFromPointer  | 421.5 ns | 4.15 ns | 3.88 ns |  0.69 |     869 B |
| XXHash64_StreamedWhole    | 423.6 ns | 1.63 ns | 1.27 ns |  0.69 |   1,123 B |
| XXHash64_StreamedMultiple | 428.6 ns | 6.49 ns | 6.07 ns |  0.70 |   1,531 B |
| XXHash64_Reference        | 611.1 ns | 2.95 ns | 2.46 ns |  1.00 |     735 B |

#### After

| Method                    | Mean     | Error   | StdDev  | Ratio | Code Size |
|-------------------------- |---------:|--------:|--------:|------:|----------:|
| XXHash64_RawPointer       | 294.3 ns | 3.89 ns | 3.64 ns |  0.48 |     667 B |
| XXHash64_FixedPointer     | 293.7 ns | 2.01 ns | 1.78 ns |  0.48 |     734 B |
| XXHash64_FixedSpan        | 292.6 ns | 0.48 ns | 0.40 ns |  0.48 |     769 B |
| XXHash64_Span             | 292.6 ns | 1.10 ns | 0.86 ns |  0.48 |     678 B |
| XXHash64_SpanFromPointer  | 291.2 ns | 1.11 ns | 0.98 ns |  0.48 |     720 B |
| XXHash64_StreamedWhole    | 328.0 ns | 0.56 ns | 0.44 ns |  0.54 |   1,030 B |
| XXHash64_StreamedMultiple | 339.3 ns | 3.79 ns | 3.55 ns |  0.56 |   1,353 B |
| XXHash64_Reference        | 608.4 ns | 0.84 ns | 0.74 ns |  1.00 |     735 B |

### Type of change

- [ ] Bug fix
- [ ] Regression bug fix
- [x] Optimization
- [ ] New feature

### How risky is the change?

- [ ] Low 
- [x] Moderate 
- [ ] High
- [ ] Not relevant

### Backward compatibility

- [x] Non breaking change
- [ ] Ensured. Please explain how has it been implemented?
- [ ] Breaking change
- [ ] Not relevant

### Is it platform specific issue?

- [x] Yes. Only for .NET 8.0 up that has `Avx512F.VL` supported
- [ ] No

### Documentation update

- [ ] This change requires a documentation update. Please mark the issue on YouTrack using `Documentation Required` tag.
- [x] No documentation update is needed 

### Testing by Contributor

- [ ] Tests have been added that prove the fix is effective or that the feature works
- [ ] Internal classes added to the test class (e.g. entity or index definition classes) have the lowest possible access modifier (preferable `private`) 
- [ ] It has been verified by manual testing
- [x] Existing tests verify the correct behavior

### Testing by RavenDB QA team

- [ ] This change requires a special QA testing due to possible performance or resources usage implications (CPU, memory, IO). Please mark the issue on YouTrack using `QA Required` tag.
- [x] No special testing by RavenDB QA team is needed

### Is there any existing behavior change of other features due to this change?

- [ ] Yes. Please list the affected features/subsystems and provide appropriate explanation
- [x] No

### UI work

- [ ] It requires further work in the Studio. Please mark the issue on YouTrack using `Studio Required` tag.
- [x] No UI work is needed